+++ Assets/Scripts/ScriptableObjects/ApplianceScriptableObjects/ApplianceBaseSO.cs
@@ -25,7 +25,7 @@ public abstract class ApplianceBaseSO : ScriptableObject
 
     public bool isTurnedOn;
 
-    public int objectWidth;
-    public int objectHeight;
-    public int objectLength;
+    public float objectWidth;
+    public float objectHeight;
+    public float objectLength;
 }
